requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.75.0
python-multipart==0.0.6
pytest==8.0.0
httpx==0.28.1
